+++ Sources/Blockus/Models/Coordinate/Coordinate+CollectionTransforms.swift
