+++ Sources/Blockus/Models/Coordinate/Coordinate.swift
