+++ Sources/Blockus/Models/Coordinate/CoordinateContainer.swift
