+++ Sources/Blockus/Models/Transform/Transform.swift
@@ -1,0 +1,13 @@
+//
+//  Transform.swift
+//  Blockus
+//
+//  Created by Chris Martin on 12/29/18.
+//
+
+import Foundation
+
+public enum Transform {
+    case mirrored(axis: Axis)
+    case rotated(amount: Rotation)
+}
